+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- sankey colors from user input and/or predefined color mapping
+- units to sankey
+
 ## [2.1.1] - 2024-06-28
 ### Fixed
 - error due to numpy version >=2.0.0